# Speedup #generate_json by using case/when/end