@byroot 2024-11-02
> * The only case where an exact class check is needed so far is for String

That's the one case where we had a failing test for it, but does that mean we should special case it?

I'm open to say only String subclasses have that behavior, but then all 3 implementations should match behavior. Could you add some tests to make sure we converge them?

Also I'm very surprised `.class` is this expensive on Truffle, I wouldn't have expected it.